Modify script to account for years with only one commit.
Add check for argument existence.
Check for case where there is only one commit in a year.
Check whether start and end commits exist before running diff.

Note that this iteration of the script assumes that if a year has only
one commit then there is a parent of that commit.  Obviously there are
cases where this will be an incorrect assumption.  See TODO in the
file.  If there is no parent of the start commit then we probably need
to count the changes in that commit toward the total number of
copyright-eligible changes.

I also think the diff over the year should probably be the diff
between the last commit of the previous year and the last commit of
the current year, in order to capture changes from the first commit of
the year.  This requires testing whether the first commit of the year
has a parent, too.