Use git diff-tree to cover the first commit of the file.
Check that the parent of the start commit exists, and behave
differently if it does or doesn't.  If it exists, then diff from that
parent (the last commit in the previous year) to the last commit in
the current year.  If it doesn't, then use diff-tree to do a diff from
the root with the --root option.

Use awk to find the insertions and deletions from git diff-tree
output (different than manually counting pluses and minuses in the
diff).