Generate MediaWiki pages from a CSV file
Add a script that makes a new MediaWiki page.  Currently this generates
a page for each line in the CSV file and fills in the content from the
cells in that row, but doesn't add any section headers.  The script
should also create a top-level page that links to all of the internal
pages it has created, but it doesn't do that yet.