Create each page just once with its correct title
Instead of renaming each page, create it after its full title is known.
This resolves problems with duplicate pages.

There is still a problem with category page counts being wrong on the
list of categories.  This may be an artifact of the testing process,
though - I've created and deleted a lot of pages in each category.  To
fix the page count for a category: click into a category page, then back
to the list of categories, then refresh.