Csv2wiki improvements for issue #1
Turn some usage comments into a usage() function.  Add full doc strings
for existing functions.  Stub out a parse_config_file function for the
new config file.  Add option parsing and turn delete into an option,
with create as the default action.

Note that the script as of this commit is BROKEN.  Use the prior version
(46bf851) or wait for the next commit.