Refactor wiki connection code into its own class
This is the followup promised in commit 3d777416a.  The new class
WikiSession inhales the config settings and does the wiki login, and
create_pages() and delete_pages() now take a WikiSession object.

Along the way, change --delete to --delete-matching=PATTERN, that is,
when doing deletion, pass the search string on the command line
instead of in the config file.