Handle page titles dynamically (from issue #3)
Along the way, encapsulate the CSV reader into the WikiSession class,
and rename the "page_title" config element to "title_tmpl".  The
change itself is not major, but the diff is large because moving the
csv reader into WikiSession meant removing one 'with ... as' loop in
create_pages(), which in turn resulted in a big re-indentation.  To
see the part of the change relevant for dynamic titles, see where the
new variables 'row_format' and 'row_num_str' are used in that function.