Require "[default]" as conf section name
The way we had been parsing the conf file, if there had been multiple
sections, then values in later sections would have overridden values
in earlier sections for fields that have the same name.  While this
shouldn't make much difference in practice (because users have no
reason to use multiple sections yet), it's still an unpredictable
behavior and violates the Principle of Least Surprise.

The solution is to consider forward compatibility and expect a
particular section, which by convention we name "default".