Specify listing suffix column in config file
There had been one bit of use-case-specific logic still in the code:
the place where we magically determined that column 1 should be used
as the suffix when listing page names in the table of contents.  Now
the run-time config file can specify which column, if any, should be
used in addition to the row number (as substituted into title_tmpl) to
help distinguish TOC entries from one another.