Make a fully generic title_tmpl template system
Implement fully generic title templating, in which arbitrary columns
can be used as components in each page title.  This eliminates the
need for 'listing_suffix_col', which is accordingly removed.

This commit finishes the title implementation, but leaves some
debugging prints in place, because Chi Hack Night is closing up now,
so I'll need to move somewhere else to finish this :-).