Make the title_tmpl template system fully generic
Arbitrary columns can now be used as components in the page title.
This eliminates the need for 'listing_suffix_col', so remove that.
The input CSV now needs to be in UTF-8 encoding, so document that.
We now depend on the 'unidecode' Python library, so document that.
Add some wiki escaping for page titles, and some for category names.
Give a better example password in the usage message.