Make the --delete-matching option work again
The --delete-matching option had been broken for a while, because we
automatically assumed the presence of a CSV input file when creating a
wiki session, even though deleting pages does not involve a CSV file.

Fixing this required a significant code rearrangement.  We now have a
new CSVInput class to encapsulate all CSV-related stuff; everything
CSV-related has been moved from WikiSession to there.  create_pages()
now takes the csv input object as an additional argument, and obtains
the header row from it.

This means create_pages() no longer obtains the header row itself, so
the code that used to do that is gone.  But I left an empty 'if False'
conditional with just 'pass' as the body, to avoid having to reindent
the entire 'else' clause and thereby obscure the semantic diff here.
I'll unwrap the conditional in a subsequent commit.