WIP: Fixing issues vs pull requests distinction
I think this code is probably wrong, but I'm pretty sure the prior
code was wrong too.  This re-separates the pull-request block from the
issue block, because they're apparently not treated as similarly as I
thought.

This discovery was driven by the realization that 'ticket.comments'
(in the old code) didn't include reviews when a ticket was a pull
request (if it ever was anyway, which I'm now not so sure of).  But
it's still not clear what 'issue.pull_request' represents, so the
conditional based on it is probably still wrong.

In any case, comment counts for pull requests from prior runs are
surely wrong, because we need to count reviews too and we weren't.