Validate authz
Twiddling the svn authz file is a process packed with pitfalls.  Today
I spent an hour trying to add a user and set some permissions.  Hooray
for terible user experiences!

On the plus side, I wrote a script to catch at least one kind of
mistake.  This script just makes sure that when you add a dir to the
authz file, it actually exists.  It lets you know if it doesn't, but
takes no other action.  It scans the current diff and the immediate
past commit on the authz file.