Teach oref-get-ref-at-point about non-nil, non-ref things
If you are so reckless as to oref-do-ref to set a ref in the middle of
a paragraph, emacs might get confused and thing you are instead trying
to dereference a ref consisting of the word under or immediately
preceding the point.  This is because thing-at-point will return a
word if it's directly after a word.  It will then prompt the user to
reindex the refs.

If, however, you are even more reckless and decide to skip oref-do-ref
and go straight to oref-set-ref, everything works as expected.

This patch just modifies oref-get-ref-at-point to handle non-blank
refs that don't match our pattern and sets them to nil.  Then
oref-do-ref sets a ref as expected.