Display total redundancy instead of ratio
The "duplication ratio" we'd been displaying since commit ec6f720fbb
was misleading: one would expect it to be the ratio of redundant file
bytes to total file bytes in the tree, but actually it was the ratio
of redundant file bytes to total bytes of all files that are part of a
duplicate set.  The latter statistic is fairly useless, and while we
could calculate the former, that's a bigger change than I want to make
right now.  So let's just display the total redundant size, and if the
users want to calculate the ratio of that to total tree size, they can
do that themselves using 'du' or whatever.