Add `ots-day-from-date' to our Emacs library
This is generally useful, but the proximate motivation is to help with
correctly dating beancount entries.  Transaction usually clear with
the bank on the next business day, so when we're adding beancount
entries based on receipts, we can usually determine the right date for
the entry if we know the day-of-week for the receipt.  But of course
the receipts don't usually say the day of the week; they just give a
numeric date.  This function allows us to quickly turn a numeric date
into a day-of-week, whence we can then deduce the next business day.

We could, if we were feeling _really_ squirrely, write a function to
return the next business day directly.  But this intermediate function
would still be necessary, and is useful in its own right, so let's
start with this and see if it's enough.